Refactor `HeaderMedia` component
- Renamed to `MainMedia`
- Tidied up file
- Updated imports across codebase
- Narrowed props to only what's needed
