+++ apps-rendering/src/capi.ts
@@ -10,8 +10,8 @@ import { TagType } from '@guardian/content-api-models/v1/tagType';
 import type { Option } from '@guardian/types';
 import { andThen, fromNullable, map, none, some } from '@guardian/types';
 import { fromString as dateFromString } from 'date';
-import type { MainMedia } from 'headerMedia';
-import { MainMediaKind } from 'headerMedia';
+import type { MainMedia } from 'mainMedia';
+import { MainMediaKind } from 'mainMedia';
 import { parseImage } from 'image';
 import { isLabs } from 'item';
 import { pipe } from 'lib';
