+++ apps-rendering/src/components/editions/headerMedia/index.tsx
@@ -12,7 +12,7 @@ import HeaderImageCaption, {
 	captionId,
 } from 'components/editions/headerImageCaption';
 import StarRating from 'components/editions/starRating';
-import { MainMediaKind } from 'headerMedia';
+import { MainMediaKind } from 'mainMedia';
 import type { Image } from 'image';
 import type { Item } from 'item';
 import { isPicture as checkIfPicture, getFormat } from 'item';
