+++ apps-rendering/src/components/layout/comment.tsx
@@ -21,7 +21,8 @@ import RelatedContent from 'components/relatedContent';
 import Series from 'components/series';
 import Standfirst from 'components/standfirst';
 import Tags from 'components/tags';
-import HeaderMedia from 'headerMedia';
+import MainMedia from 'mainMedia';
+import { getFormat } from 'item';
 import type { Comment as CommentItem, Editorial, Letter } from 'item';
 import type { FC, ReactNode } from 'react';
 import {
@@ -99,7 +100,10 @@ const Comment: FC<Props> = ({ item, children }) => (
 					<Metadata item={item} />
 				</section>
 
-				<HeaderMedia item={item} />
+				<MainMedia
+					format={getFormat(item)}
+					mainMedia={item.mainMedia}
+				/>
 				<section css={articleWidthStyles}>
 					<Logo item={item} />
 				</section>
