+++ apps-rendering/src/components/layout/interactiveimmersive.tsx
@@ -19,7 +19,8 @@ import RelatedContent from 'components/relatedContent';
 import Series from 'components/series';
 import Standfirst from 'components/standfirst';
 import Tags from 'components/tags';
-import HeaderMedia from 'headerMedia';
+import MainMedia from 'mainMedia';
+import { getFormat } from 'item';
 import type {
 	MatchReport as MatchReportItem,
 	Review as ReviewItem,
@@ -70,7 +71,10 @@ const InteractiveImmersive: FC<Props> = ({ item, children }) => {
 		<main css={[Styles, DarkStyles]}>
 			<article className="js-article" css={BorderStyles}>
 				<header>
-					<HeaderMedia item={item} />
+					<MainMedia
+						format={getFormat(item)}
+						mainMedia={item.mainMedia}
+					/>
 					<Series item={item} />
 					<Headline item={item} />
 					<div css={articleWidthStyles}>
